conf/requirements.txt
python-dotenv
requests
csv
